requirements.txt
openai>=1.40.0
jira>=3.8.0
python-dotenv>=1.0.1
rich>=13.7.0
